simple tcp echo server with test Test is absurdly complex at the moment.  I'm playing around with contexts, connections, error groups, and wait groups to ensure that the test both observes a time limit and that all goroutines end promptly upon error.
Of course... I don't at this point test what happens on an error.
I can do that next for fun, and then try to simplify this a bit.
I do think it does a nice job of illustrating how a passed context is
used to cancel blocking actions (using AfterFunc to close connection)