Non-async io TCP server/client (Rust implementation) (#1)
Initial commit for a simple TCP rust server / client. The client will write with the server waiting, and then the server will echo back to the client. This all happens in a tight loop with multiple threads running for each. There are a couple issues here that will be addressed in later commits: 
 - The number of clients should be set via command line argument. Will find a library for handling command line arguments.
 - Switching to Unix domain socket rather than TCP
 - Reading until a delimiter and then begin writing, rather than reading until the transmission is complete