rust async WIP
new goal is to set up socket/pipe/futex sync and async implementations that are
close to identical as possible.  Even the syscalls should be identical except
that sync will block/context switch in the syscall whereas async will only allow
a task to continue once it is known that the syscall (e.g., socket read) will
not block.  We can either use NOWAIT and assert that we get data or use a
blocking syscall and then monitor context switches.

We can then compare performance and look at:
- instructions per iteration
- IPC of each.  If IPC is similar then the cost of a context switch is direct
(running instructions), if sync IPC is lower then the cost is indirect.
- TLB and cache misses in support of indirect cost.

We can also vary the number of threads/tasks or add in a compute heavy work with
controlled memory stride to examine cache performance.